Refactor: MemberType 공유 객체에서 관리
