+++ gathering-service/src/main/java/com/sparta/moim/gathering/gathering/application/service/rds/MemberServiceAsRds.java
@@ -7,9 +7,9 @@ import com.sparta.moim.gathering.gathering.application.exception.AlreadyParticip
 import com.sparta.moim.gathering.gathering.application.exception.NotOpenGatheringException;
 import com.sparta.moim.gathering.gathering.application.exception.RoleNotAllowedGatheringException;
 import com.sparta.moim.gathering.gathering.application.service.MemberService;
-import com.sparta.moim.gathering.gathering.domain.enums.MemberType;
 import com.sparta.moim.gathering.gathering.domain.repository.GatheringValidationRepository;
 import com.sparta.moim.gathering.gathering.domain.repository.MemberRepository;
+import com.sparta.moim.gathering.sherd.enums.MemberType;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
