+++ gathering-service/src/main/java/com/sparta/moim/gathering/gathering/domain/entity/Member.java
@@ -1,7 +1,8 @@
 package com.sparta.moim.gathering.gathering.domain.entity;
 
-import com.sparta.moim.gathering.gathering.domain.enums.MemberType;
+
 import com.sparta.moim.gathering.gathering.application.dto.event.GatheringAddAdminEvent;
+import com.sparta.moim.gathering.sherd.enums.MemberType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
