--- gathering-service/src/main/java/com/sparta/moim/gathering/gathering/domain/enums/MemberType.java
@@ -1,5 +1,0 @@
-package com.sparta.moim.gathering.gathering.domain.enums;
-
-public enum MemberType {
-  ADMIN, GENERAL
-}
