+++ gathering-service/src/main/java/com/sparta/moim/gathering/sherd/enums/MemberType.java
@@ -1,0 +1,5 @@
+package com.sparta.moim.gathering.sherd.enums;
+
+public enum MemberType {
+  ADMIN, GENERAL
+}
